tradino_unschlagbar/requirements.txt
# requirements.txt - Kompatible Dependencies für Python 3.12

# Core Trading Dependencies
loguru>=0.7.3
pydantic>=2.11.7
aiohttp>=3.12.13
pandas>=2.3.0
numpy>=2.3.1
ccxt>=4.4.90
python-telegram-bot>=22.1
scikit-learn>=1.7.0
xgboost>=3.0.2
ta>=0.11.0
python-dotenv>=1.1.0
PyYAML>=6.0.2
psutil>=7.0.0

# RL Dependencies - Reinforcement Learning System
torch>=2.0.0
stable-baselines3>=2.0.0
gym>=0.26.0
shimmy>=2.0.0
scipy>=1.9.0
tensorboard>=2.12.0
matplotlib>=3.5.0 